1. Bug fixing for string dictionary collect handling null value for multi-value column. 2. Bug fixing for LongDictionary Reader.
RB=465583
R=xiafu
A=dpatel

+++ pinot-core/src/main/java/com/linkedin/pinot/core/segment/index/readers/LongDictionary.java
@@ -70,7 +70,7 @@ public class LongDictionary extends ImmutableDictionaryReader {
   }
 
   private long getLong(int dictionaryId) {
-    return dataFileReader.getInt(dictionaryId, 0);
+    return dataFileReader.getLong(dictionaryId, 0);
   }
 
 }
